requirements.txt
flask
flask_jwt_extended
flask_limiter
pymongo
kafka-python
cryptography